// backup11-07/app/common/interfaces.ts
export interface IWidgetInfo {
    name: string;
    title: string;
    dataProvider?: IActionInfo;
    metaType?: string; // TODO: change string to pre-define type 
    transformationType?: string;
    widgetConfig?: IWidgetConfig;
    metaConfig?: any;
}

export interface IActionInfoParent {
    title?: string;
}

export interface IActionAferCall {
    type: string;
    widgetInfoList: IWidgetActionAfterResponce[];
    urlToNavigate?: string;
    actionInfo?: IActionInfo;
    hideSource?: boolean;
}

export interface IWidgetActionAfterResponce {
    widget?: IWidgetInfo;
    widgetKey?: string;
    dataAction: string;
    fieldId?: string;
}

export interface IActionInfo extends IActionInfoParent {
    id?: string;
    type: string;
    method?: string;
    context?: string;
    url?: string;
    dev_url?: string;
    parameters?: IActionParameter[];
    data?: any;
    transformationType?: string;
    otherDetails?: any;
    sendAllParam?: boolean;
    responseHandler?: IActionAferCall;
    errorHandler?:IActionAferCall;
    parameterValueType?: string;
    doFileUpload?: boolean;
    fileUploadFields?: any[];
    fileUploadAction?: IActionInfo;
}
// ...
export interface IWidgetToggleSettings {
    label?: string;
    imageUrl?: string;
    widgetInfo: IWidgetInfo;
    widgetConfig: IWidgetConfig;
}

export interface IElementPosition {
    left: number;
    top: number;
    height?: number;
    width?: number;
}

export interface IWidgetConfig {
    showHeader?: boolean;
    isPlainWidget?: boolean;
    customClass?: string;
}